// src/components/forms/input-profile/input-base/enum.ts
enum TagNames {
  DIV = 'div',
  LABEL = 'label',
  INPUT = 'input',
}

enum Styles {
  INPUT_CONTAINER = 'input-container',
  LABEL = 'label',
  INPUT = 'input-profile',
  INPUT_ERROR = 'input--error',
}

enum Attributes {
  ID = 'id',
  ID_VALUE_INPUT = 'email',
  TYPE = 'type',
  FOR = 'for',
  NAME = 'name',
  NAME_VALUE = 'email',
  PLACEHOLDER = 'placeholder',
  PLACEHOLDER_VALUE = 'Enter email',
}

enum Contents {
  LABEL = 'Email',
}

enum Events {
  INPUT = 'input',
}

export { TagNames, Styles, Events, Contents, Attributes };